node-red: update npm-shrinkwrap
the previous shrinkwrap was already for the current node-red version (v2.0.5),
but it was made using an older npm version. With this update, its all up to date
and generated with the npm version shipped with nodejs v14, which we now use.

Besides some changes in outline, also some dependencies have (inevitably) been
updated to a newer version, as that happens almost everytime when doing a new
install.