node-red-contrib-victron: update npm-shrinkwrap
The prior version of the shrinkwrap was made for node-red-contrib-victron v1.3.2.
Since then, when increasing the installed signalk-server version, the shrinkwrap
was never refreshed. The shrinkwrap specifies which version is installed for each
of the dependencies. So, it being outdated, meant that many of the installed
dependencies were outdated as well.

Surprisingly, having those old dependencies didn't cause any noticable issues for
users.

Going forward, everytime we increase the version of a nodejs package, also its
shrinkwrap must be refreshed.